Move the graph-building methods to `ReferenceTerm`
