+++ analyzer/src/main/kotlin/com/github/cubuspl42/sigmaLang/analyzer/semantics/expressions/Call.kt
@@ -79,7 +79,7 @@ abstract class Call : FirstOrderExpression() {
                     term = term.rightArgument,
                 ).resolved
 
-                val subjectStub = Reference.build(
+                val subjectStub = ReferenceTerm.build(
                     context,
                     referredName = Symbol.of(prototype.functionName),
                 )
